+++ royal/tests/functional/test_example.py
@@ -171,3 +171,7 @@ class Test(TestBase):
             },
         }
         self.assertEqual(expected, result.json)
+
+    def test_post_item_returns_405(self):
+        user = self._add_single_user('franky', 'franky@email.com')
+        self.app.post('/users/franky', {'key': 'value'}, status=405)