+++ royal/interfaces.py
@@ -44,12 +44,6 @@ class ICollection(IBase):
         "DELETE /items"
 
 
-class IResourceConfigurator(Interface):
-
-    def add_resource(dot_path):
-        "Add a resource by its dot notation: 'apps.users' is apps/{id}/users"
-
-
 class IDeserializer(Interface):
 
     def add_deserializer(content_type, callable):